# (Performance) improvements on gathering field information from `MixedDofHandler`