--- a/docs/src/reference/dofhandler.md
+++ b/docs/src/reference/dofhandler.md
@@ -34,16 +34,6 @@ celldofs
 celldofs!
 ```
 
-## `AbstractDofHandler` interface
-The following are *internal* methods that need to be implemented to fulfill the `AbstractDofHandler`
-interface.
-```@docs
-Ferrite.find_field
-Ferrite.field_offset
-Ferrite.getfieldinterpolation
-Ferrite.getfielddim
-```
-
 # CellIterator
 ```@docs
 CellCache