This PR tackles the functionality that requests field information from the `MixedDofHandler`, in particular 
- `find_field`
- `getfielddim`
- `getfieldinterpolation`
- `field_offset`
- `dof_range`

## Performance improvements
Finding fields based on the field name was a bit suboptimal. From the side of the user API, in particular `dof_range` profits, but most improvements are based on a better `find_field`:
|    | #master | #PR |
|---|---|---|
| `@btime dof_range($fh, $(:f3))` | 567.935 ns (14 allocations: 672 bytes) | 36.715 ns (0 allocations: 0 bytes)|
| `@btime Ferrite.find_field($mixed_dh, $(:f3))`| 31.448 ns (1 allocation: 80 bytes) | 4.583 ns (0 allocations: 0 bytes) |
| `@btime Ferrite.find_field($fh, $(:f3))` | 29.104 ns (1 allocation: 80 bytes) | 4.583 ns (0 allocations: 0 bytes) |

<details>
<summary>Benchmark code</summary>

```julia
using Ferrite
using BenchmarkTools

grid = generate_grid(Quadrilateral, (10, 10))

# three fields, each on all cells
dh = DofHandler(grid)
add!(dh, :f1, 1, Lagrange{2,RefCube,1}())
add!(dh, :f2, 2, Lagrange{2,RefCube,1}())
add!(dh, :f3, 1, Lagrange{2,RefCube,2}())
close!(dh)

mixed_dh = MixedDofHandler(grid)
add!(mixed_dh, :f1, 1, Lagrange{2,RefCube,1}())
add!(mixed_dh, :f2, 2, Lagrange{2,RefCube,1}())
add!(mixed_dh, :f3, 1, Lagrange{2,RefCube,2}())
close!(mixed_dh)

fh = mixed_dh.fieldhandlers[1]

@btime dof_range($fh, $(:f3))

@btime Ferrite.find_field($mixed_dh, $(:f3))
@btime Ferrite.find_field($fh, $(:f3))

```
</details>

###  Breaking change to `find_field(mixed_dh, fieldname)`
```julia
# master:
julia> Ferrite.find_field(mixed_dh, :f3)
3

# PR
julia> Ferrite.find_field(mixed_dh, :f3)
(1, 3)
```
`find_field` now searches for the first occurence of the given field and returns a tuple with the index of the fieldhandler it was found in and the index of the field in the fieldhandler (before it only searched in the first fieldhandler).

## API
It is most efficient to call `getfielddim`, `getfieldinterpolation`, `field_offset` and `dof_range` with `field_idxs` instead of `fieldname` as argument. However, in order to have the exact same syntax for `MixedDofHandler` and `DofHandler` (working towards merging them), I suggest to have both options for all functions, i.e. possible ways to call e.g. `dof_range` are:
- `dof_range(mixed_dh, field_idxs)`
- `dof_range(mixed_dh, fieldname)`
- `dof_range(fh, field_idxs)`
- `dof_range(fh, fieldname)`