@kimauth 2023-03-20
It's not broken, that is the intended behaviour. (It's described in the docstring like that.) 

However, after some discussion with @fredrikekre, I'll add error messages to some of the functions for the case that a `MixedDofHandler` has several `FieldHandlers`.