@termi-official 2023-03-20
Okay. Then I have a suggestion for the dof handler merge. Do you think it makes sense to have different FieldHandlers? One for fields covering the full domain (i.e. the functionality of the current dof handler) and one for fields on subdomains (which is the selling point of the mixed dof handler)? cc @fredrikekre