@fredrikekre 2023-03-20
That is what Kim and I discussed. We rename `MixedDofHandler` to `DofHandler` (removing the old one), and rename `FieldHandler` to `SubDofHandler`. If you are using the subdomain functionality, then you are forced to use the corresponding `SubDofHandler`s, but if you don't, then you can request things from the `DofHandler` directly. Old `DofHandler` would then just be new `DofHandler` with a single `SubDofHandler`, that you don't need to worry about.