@koehlerson 2023-03-20
How should I know then externally if a field is defined on all types of cells as e.g. `:v` in this example: https://github.com/Ferrite-FEM/Ferrite.jl/blob/master/test/test_mixeddofhandler.jl#L437-L485 ? Do I have to loop then in any case over `SubDofHandler`s?